requirements.txt
GitPython==3.0.5
GPUtil==1.4.0
PyMySQL==0.9.3